+++ packages/integration-sdk-runtime/src/synchronization/index.ts
@@ -42,6 +42,7 @@ export interface SynchronizeInput {
   integrationInstanceId: string;
   integrationJobId?: string;
   uploadBatchSize?: number | undefined;
+  uploadRelationshipBatchSize?: number | undefined;
 }
 
 /**
@@ -88,6 +89,7 @@ export interface SynchronizationJobContext {
   job: SynchronizationJob;
   logger: IntegrationLogger;
   uploadBatchSize?: number | undefined;
+  uploadRelationshipBatchSize?: number | undefined;
 }
 
 /**
@@ -99,6 +101,7 @@ export async function initiateSynchronization({
   integrationInstanceId,
   integrationJobId,
   uploadBatchSize,
+  uploadRelationshipBatchSize,
 }: SynchronizeInput): Promise<SynchronizationJobContext> {
   logger.info('Initiating synchronization job...');
 
@@ -127,6 +130,7 @@ export async function initiateSynchronization({
       integrationInstanceId: job.integrationInstanceId,
     }),
     uploadBatchSize,
+    uploadRelationshipBatchSize,
   };
 }
 
@@ -183,7 +187,12 @@ export async function uploadGraphObjectData(
   synchronizationJobContext: SynchronizationJobContext,
   graphObjectData: FlushedGraphObjectData,
   uploadBatchSize?: number,
+  uploadRelationshipsBatchSize?: number,
 ) {
+  const entityBatchSize = uploadBatchSize;
+  const relationshipsBatchSize =
+    uploadRelationshipsBatchSize || uploadBatchSize;
+
   try {
     if (
       Array.isArray(graphObjectData.entities) &&
@@ -200,7 +209,7 @@ export async function uploadGraphObjectData(
         synchronizationJobContext,
         'entities',
         graphObjectData.entities,
-        uploadBatchSize,
+        entityBatchSize,
       );
 
       synchronizationJobContext.logger.info(
@@ -226,7 +235,7 @@ export async function uploadGraphObjectData(
         synchronizationJobContext,
         'relationships',
         graphObjectData.relationships,
-        uploadBatchSize,
+        relationshipsBatchSize,
       );
 
       synchronizationJobContext.logger.info(
@@ -248,7 +257,12 @@ export async function uploadCollectedData(context: SynchronizationJobContext) {
   context.logger.synchronizationUploadStart(context.job);
 
   async function uploadGraphObjectFile(parsedData: FlushedGraphObjectData) {
-    await uploadGraphObjectData(context, parsedData, context.uploadBatchSize);
+    await uploadGraphObjectData(
+      context,
+      parsedData,
+      context.uploadBatchSize,
+      context.uploadRelationshipBatchSize,
+    );
   }
 
   await timeOperation({